% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{addTrees}
\alias{addTrees}
\title{addTrees-forestry}
\usage{
addTrees(object, ntree)
}
\arguments{
\item{object}{A `forestry` object.}

\item{ntree}{Number of new trees to add}
}
\value{
A `forestry` object
}
\description{
Add more trees to the existing forest.
}
