% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwards_compatible.R
\name{autohonestRF}
\alias{autohonestRF}
\title{Honest Random Forest}
\usage{
autohonestRF(...)
}
\arguments{
\item{...}{parameters which are passed directly to `autoforestry`}
}
\description{
This function is deprecated and only exists for backwards
  backwards compatibility. The function you want to use is `autoforestry`.
}
