% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{bcPredict}
\alias{bcPredict}
\title{bcPredict-forestry}
\usage{
bcPredict(object, newdata, feats, nrounds, linear)
}
\arguments{
\item{object}{A `forestry` object.}

\item{newdata}{Dataframe on which to predict.}

\item{feats}{A vector of feature indices which should be included in the bias
correction. By default only the outcome and predicted outcomes are used.}

\item{nrounds}{The number of nonlinear bias correction steps which should be
taken. By default this is zero, so just a single linear correction is used.}

\item{linear}{A flag indicating whether or not we want to do a final linear
bias correction after doing the nonlinear corrections. Default is TRUE.}
}
\value{
A vector of the bias corrected predictions
}
\description{
Perform predictions given the forest using a bias correction based on
  the out of bag predictions on the training set. By default we use a final linear
  correction based on the leave-one-out hat matrix after doing `nrounds` nonlinear
  corrections.
}
