% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{correctedPredict}
\alias{correctedPredict}
\title{correctedPredict-forestry}
\usage{
correctedPredict(
  object,
  newdata,
  feats = NULL,
  nrounds = 0,
  linear = TRUE,
  double = FALSE,
  simple = FALSE,
  verbose = FALSE,
  use_residuals = FALSE,
  adaptive = FALSE,
  monotone = FALSE,
  num_quants = 5
)
}
\arguments{
\item{object}{A `forestry` object.}

\item{newdata}{Dataframe on which to predict.}

\item{feats}{A vector of feature indices which should be included in the bias
correction. By default only the outcome and predicted outcomes are used.}

\item{nrounds}{The number of nonlinear bias correction steps which should be
taken. By default this is zero, so just a single linear correction is used.}

\item{linear}{A flag indicating whether or not we want to do a final linear
bias correction after doing the nonlinear corrections. Default is TRUE.}

\item{double}{flag indicating if one should use doubleOOB for the initial
predictions rather than "oob." Defalt is FALSE.}

\item{simple}{flag indicating whether we should do a simple linear adjustment
or do different adjustments by quantiles. Default is FALSE.}

\item{verbose}{Flag which displays the bias of each qunatile.}

\item{use_residuals}{flag indicating if we should use the residuals to fit the
bias correction steps. Defualt is FALSE which means that we will use Y
rather than Y-Y.hat as the regression outcome in the bias correction steps.}

\item{adaptive}{flag to indicate whether we use adaptiveForestry or not in the
regression step.}

\item{monotone}{flag to indicate whether or not we should use monotonicity
in the regression of Y on Y hat (when doing forest correction steps).
If TRUE, will constrain the corrected prediction for Y to be monotone in the
original prediction of Y. Default is FALSE.}

\item{num_quants}{Number of quantiles to use when doing quantile specific bias
correction.}
}
\value{
A vector of the bias corrected predictions
}
\description{
Perform predictions given the forest using a bias correction based on
  the out of bag predictions on the training set. By default we use a final linear
  correction based on the leave-one-out hat matrix after doing `nrounds` nonlinear
  corrections.
}
