% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{export_json}
\alias{export_json}
\title{export_json}
\usage{
export_json(object)
}
\arguments{
\item{object}{A `forestry` object.}
}
\value{
A string containing Treelite JSON of the forest
}
\description{
Returns a JSON string representation of the forest. The JSON format aligns with the spec defined for import by the Treelite library.
}
