% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{export_treelite_json}
\alias{export_treelite_json}
\title{export_treelite_json}
\usage{
export_treelite_json(object)
}
\arguments{
\item{object}{A `forestry` object.}
}
\value{
A string containing Treelite JSON of the forest
}
\description{
Return a string containing Treelite JSON of the forest https://treelite.readthedocs.io/en/latest/tutorials/json_import.html
}
