% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{forest_checker}
\alias{forest_checker}
\title{Checks if forestry object has valid pointer for C++ object.}
\usage{
forest_checker(object)
}
\arguments{
\item{object}{a forestry object}
}
\description{
Checks if forestry object has valid pointer for C++ object.
}
