% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{getSplitProps-forestry}
\alias{getSplitProps-forestry}
\alias{getSplitProps}
\title{getSplitProps-forestry}
\usage{
getSplitProps(object)
}
\arguments{
\item{object}{A trained model object of class "forestry".}
}
\value{
A vector of length equal to the number of columns
}
\description{
Retrieves the proportion of splits for each feature in the given
 forestry object. These proportions are calculated as the number of splits
 on feature i in the entire forest over total the number of splits in the
 forest.
}
\seealso{
\code{\link{forestry}}
}
