% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{getVI}
\alias{getVI}
\title{getVI-forestry}
\usage{
getVI(object, noWarning, seed = NULL)
}
\arguments{
\item{object}{A `forestry` object.}

\item{noWarning}{flag to not display warnings}

\item{seed}{A parameter to seed the random number generator for shuffling
the features of X.}
}
\value{
The variable importance of the forest.
}
\description{
Calculate the percentage increase in OOB error of the forest
 when each feature is shuffled.
}
\note{
No seed is passed to this function so it is
  not possible in the current implementation to replicate the vector
  permutations used when measuring feature importance.
}
