% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwards_compatible.R
\name{honestRF}
\alias{honestRF}
\title{Honest Random Forest}
\usage{
honestRF(...)
}
\arguments{
\item{...}{parameters which are passed directly to `forestry`}
}
\description{
This function is deprecated and only exists for backwards
  backwards compatibility. The function you want to use is `forestry`.
}
