% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{predict-multilayer-forestry}
\alias{predict-multilayer-forestry}
\alias{predict.multilayerForestry}
\title{predict-multilayer-forestry}
\usage{
\method{predict}{multilayerForestry}(
  object,
  feature.new,
  aggregation = "average",
  seed = as.integer(runif(1) * 10000),
  ...
)
}
\arguments{
\item{object}{A `multilayerForestry` object.}

\item{feature.new}{A data frame of testing predictors.}

\item{aggregation}{How shall the leaf be aggregated. The default is to return
the mean of the leave `average`. Other options are `weightMatrix`.}

\item{seed}{random seed}

\item{...}{additional arguments.}
}
\value{
A vector of predicted responses.
}
\description{
Return the prediction from the forest.
}
