% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{relinkCPP_prt}
\alias{relinkCPP_prt}
\title{relink CPP ptr}
\usage{
relinkCPP_prt(object)
}
\arguments{
\item{object}{an object of class `forestry` or class `multilayerForestry`}
}
\description{
When a `foresty` object is saved and then reloaded the Cpp
  pointers for the data set and the Cpp forest have to be reconstructed
}
