% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_preprocessing.R
\name{scale_center}
\alias{scale_center}
\title{scale_center}
\usage{
scale_center(x, categoricalFeatureCols, colMeans, colSd)
}
\arguments{
\item{x}{A dataframe in order to be processed.}

\item{categoricalFeatureCols}{A vector of the categorical features, we
don't want to scale/center these. Should be 1-indexed.}

\item{colMeans}{A vector of the means to center each column.}

\item{colSd}{A vector of the standard deviations to scale each column with.}
}
\value{
A scaled and centered  dataset x
}
\description{
Given a dataframe, scale and center the continous features
}
