% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_preprocessing.R
\name{unscale_uncenter}
\alias{unscale_uncenter}
\title{unscale_uncenter}
\usage{
unscale_uncenter(x, categoricalFeatureCols, colMeans, colSd)
}
\arguments{
\item{x}{A dataframe in order to be processed.}

\item{categoricalFeatureCols}{A vector of the categorical features, we
don't want to scale/center these. Should be 1-indexed.}

\item{colMeans}{A vector of the means to add to each column.}

\item{colSd}{A vector of the standard deviations to rescale each column with.}
}
\value{
A dataset x in it's original scaling
}
\description{
Given a dataframe, un scale and un center the continous features
}
