% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{getOOB-forestry}
\alias{getOOB-forestry}
\alias{getOOB}
\alias{getOOB,forestry-method}
\title{getOOB-forestry}
\usage{
getOOB(object, noWarning)
}
\arguments{
\item{object}{A `forestry` object.}

\item{noWarning}{flag to not display warnings}
}
\value{
The OOB error of the forest.
}
\description{
Calculate the out-of-bag error of a given forest.
}
