% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry.R
\name{getOOBpreds-forestry}
\alias{getOOBpreds-forestry}
\alias{getOOBpreds}
\title{getOOBpreds-forestry}
\usage{
getOOBpreds(object, noWarning)
}
\arguments{
\item{object}{A trained model object of class "forestry".}

\item{noWarning}{Flag to not display warnings.}
}
\value{
The vector of all training observations, with their out of bag
 predictions. Note each observation is out of bag for different trees, and so
 the predictions will be more or less stable based on the observation. Some
 observations may not be out of bag for any trees, and here the predictions
 are returned as NA.
}
\description{
Calculate the out-of-bag predictions of a given forest.
}
\seealso{
\code{\link{forestry}}
}
